Move RequestHandler to common/server, eliminating duplication between source and dev/app (#570)
