+++ dev/app/lib/handlers/ContentHandler.js
@@ -1,5 +1,5 @@
 import { notFound } from './not_found.js';
-import RequestHandler from './RequestHandler.js';
+import { RequestHandler } from '../common/server/RequestHandler.js';
 import { Logger } from '../common/utils/logging/Logger.js';
 import DataNavigator from '../models/DataNavigator.js';
 import RouteParamsExtractor from '../routing/RouteParamsExtractor.js';
