+++ source/lib/server/handlers/StatsRequestHandler.js
@@ -1,6 +1,6 @@
 import { JobRegistry } from '../../background/JobRegistry.js';
 import { WorkersRegistry } from '../../background/WorkersRegistry.js';
-import { RequestHandler } from '../RequestHandler.js';
+import { RequestHandler } from '../../common/server/RequestHandler.js';
 
 /**
  * Handles GET /stats.json requests.
