+++ source/lib/server/handlers/engine/EnginePauseRequestHandler.js
@@ -1,6 +1,6 @@
+import { RequestHandler } from '../../../common/server/RequestHandler.js';
 import { ConflictError } from '../../../exceptions/http/ConflictError.js';
 import { Application } from '../../../services/Application.js';
-import { RequestHandler } from '../../RequestHandler.js';
 
 /**
  * Handles PATCH /engine/pause requests.
