+++ source/lib/server/handlers/engine/EngineShutdownRequestHandler.js
@@ -1,5 +1,5 @@
+import { RequestHandler } from '../../../common/server/RequestHandler.js';
 import { Application } from '../../../services/Application.js';
-import { RequestHandler } from '../../RequestHandler.js';
 
 /**
  * Handles PATCH /engine/shutdown requests.
