+++ source/lib/server/handlers/jobs/JobRetryRequestHandler.js
@@ -1,7 +1,7 @@
 import { JobRegistry } from '../../../background/JobRegistry.js';
+import { RequestHandler } from '../../../common/server/RequestHandler.js';
 import { ConflictError } from '../../../exceptions/http/ConflictError.js';
 import { NotFoundError } from '../../../exceptions/http/NotFoundError.js';
-import { RequestHandler } from '../../RequestHandler.js';
 
 const RETRYABLE_STATUSES = new Set(['failed', 'dead']);
