--- source/spec/lib/server/RequestHandler_spec.js
@@ -1,10 +1,0 @@
-import { RequestHandler } from '../../../lib/server/RequestHandler.js';
-
-describe('RequestHandler', () => {
-  describe('#handle', () => {
-    it('is a no-op', () => {
-      const handler = new RequestHandler();
-      expect(() => handler.handle({}, {})).not.toThrow();
-    });
-  });
-});
